use MsalAuthenticationTemplate
