+++ src/App.component.tsx
@@ -5,8 +5,6 @@ import MenuComponent from "features/menu/Menu.component";
 import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
-import { Configuration, PublicClientApplication } from "@azure/msal-browser";
-import { MsalProvider } from "@azure/msal-react";
 import RightPanelComponent from "features/right-panel/RightPanel.component";
 import ContentComponent from "features/app-content/Content.component";
 import OverviewComponent from "features/overview/Overview.component";
@@ -18,7 +16,6 @@ import AboutComponent from "features/about/About.component";
 import TestingStrategiesComponent from "features/testing-strategies/TestingStrategies.component";
 import AccountPerformanceComponent from "features/account-performance/AccountPerformance.component";
 import MetricsComponent from "features/metrics/Metrics.component";
-import AuthenticationContainerComponent from "features/authentication/AuthenticationContainer.component";
 import RequestAuthenticationInterceptorComponent from "features/authentication/RequestAuthenticationInterceptor.component";
 
 type AppProps = AppReduxProps & AppDispatchProps;
@@ -31,16 +28,6 @@ interface AppReduxProps
 {
     darkMode: boolean;
 }
-
-const config: Configuration = {
-    auth: {
-        clientId: "96a42910-ace2-4e54-bcd5-b34e0275cab0",
-        authority: "https://login.microsoftonline.com/09c105c2-92b3-4242-aa95-062b36c2534c",
-        redirectUri: "http://localhost:3000"
-    }
-};
-
-const msalInstance = new PublicClientApplication(config);
 
 export const AppComponent = ({ darkMode, initializePreferences }: AppProps) => {
     const [tabMap, setTabMap] = useState<Map<string, [string, JSX.Element]>>(new Map<string, [string, JSX.Element]>());
@@ -70,23 +57,19 @@ export const AppComponent = ({ darkMode, initializePreferences }: AppProps) => {
     }, [darkMode]);
 
     return (
-        <MsalProvider instance={msalInstance!}>
-            <AuthenticationContainerComponent>
-                <RequestAuthenticationInterceptorComponent>
-                    <Container fluid className="p-0">
-                        <Row>
-                            <MenuComponent tabMap={tabMap}></MenuComponent>
-                            <Col xs={9}>
-                                <ContentComponent tabMap={tabMap}></ContentComponent>
-                            </Col>
-                            <Col xs={3}>
-                                <RightPanelComponent></RightPanelComponent>
-                            </Col>
-                        </Row>
-                    </Container>
-                </RequestAuthenticationInterceptorComponent>
-            </AuthenticationContainerComponent>
-        </MsalProvider>
+        <RequestAuthenticationInterceptorComponent>
+            <Container fluid className="p-0">
+                <Row>
+                    <MenuComponent tabMap={tabMap}></MenuComponent>
+                    <Col xs={9}>
+                        <ContentComponent tabMap={tabMap}></ContentComponent>
+                    </Col>
+                    <Col xs={3}>
+                        <RightPanelComponent></RightPanelComponent>
+                    </Col>
+                </Row>
+            </Container>
+        </RequestAuthenticationInterceptorComponent>
     );
 }
